Alterado nome de Notification para AlertNotification
