+++ grails-app/views/alertNotification/index.gsp
@@ -10,24 +10,24 @@
     <body>
         <div class="container pt-5 pb-3" style="max-width: 576px">
             <h1 class="mb-4">Notificações</h1>
-            <g:if test="${notificationList.isEmpty()}">
+            <g:if test="${alertNotificationList.isEmpty()}">
                 <div class="card card-body">
                     <p class="card-text text-center">Você não possui notificações</p>
                 </div>
             </g:if>
             <g:else>
-                <g:each var="notification" in="${notificationList}">
-                    <g:set var="notificationTypeCode" value="NotificationType.${notification.type}"/>
+                <g:each var="alertNotification" in="${alertNotificationList}">
+                    <g:set var="alertNotificationTypeCode" value="AlertNotificationType.${alertNotification.type}"/>
                     <div class="card card-body p-0 mb-3">
-                        <g:link action="show" id="${notification.publicId}" class="text-decoration-none p-3 d-block ${!notification.unread ? 'read' : ''}" style="color: inherit">
+                        <g:link action="show" id="${alertNotification.publicId}" class="text-decoration-none p-3 d-block ${!alertNotification.unread ? 'read' : ''}" style="color: inherit">
                             <p class="fw-bold mb-1">
-                                <g:message code="${notificationTypeCode}.title"/>
+                                <g:message code="${alertNotificationTypeCode}.title"/>
                             </p>
                             <p class="mb-1">
-                                <g:message code="${notificationTypeCode}.content" args="[notification.payment.payer.name, currencyFormat([value: notification.payment.value])]"/>
+                                <g:message code="${alertNotificationTypeCode}.content" args="[alertNotification.payment.payer.name, currencyFormat([value: alertNotification.payment.value])]"/>
                             </p>
                             <p class="text-muted mb-0">
-                                <g:datetimeFormat value="${notification.createdAt}"/>
+                                <g:datetimeFormat value="${alertNotification.createdAt}"/>
                             </p>
                         </g:link>
                     </div>
